# Conflicts:
#	backend/src/main/java/com/anouar/grabit/repository/CourierRepository.java
#	backend/src/main/java/com/anouar/grabit/service/CourierService.java
#	backend/src/main/java/com/anouar/grabit/service/CourierServiceImpl.java

+++ backend/src/main/java/com/anouar/grabit/repository/CourierRepository.java
@@ -4,9 +4,8 @@ import com.anouar.grabit.model.Courier;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-
 @Repository
-public interface CourierRepository extends JpaRepository<Courier, Long> {
+public interface CourierRepository extends JpaRepository<Courier, Integer> {
 
     Courier findByEmail(String email);
